Split ConnectionBase into connection_base.py
