+++ randovania/game_connection/game_connection.py
@@ -3,7 +3,8 @@ from typing import List
 from PySide2.QtCore import QTimer, Signal, QObject
 from asyncqt import asyncSlot
 
-from randovania.game_connection.connection_backend import ConnectionBackend, ConnectionBase, ConnectionStatus
+from randovania.game_connection.connection_backend import ConnectionBackend
+from randovania.game_connection.connection_base import ConnectionStatus, ConnectionBase
 from randovania.game_description.resources.pickup_entry import PickupEntry
 from randovania.game_description.resources.resource_info import CurrentResources
