+++ test/gui/test_debug_backend_window.py
@@ -2,7 +2,7 @@ import pytest
 from PySide2.QtCore import Qt
 from mock import patch, MagicMock, AsyncMock
 
-from randovania.game_connection.connection_backend import ConnectionStatus
+from randovania.game_connection.connection_base import ConnectionStatus
 from randovania.game_description.item.item_category import ItemCategory
 from randovania.game_description.resources.pickup_entry import PickupEntry, ConditionalResources
 from randovania.gui.debug_backend_window import DebugBackendWindow
