+++ tools/game_console.py
@@ -5,7 +5,7 @@ import sys
 
 from PySide2.QtCore import QCoreApplication
 
-from randovania.game_connection.connection_backend import ConnectionStatus
+from randovania.game_connection.connection_base import ConnectionStatus
 from randovania.game_connection.dolphin_backend import DolphinBackend
 from randovania.game_connection.game_connection import GameConnection
